Add functionality to calculator
Added functionality so that pressing buttons populates the display and
calculates the expression.

Right now, division by zero is correctly caught, but not dealt with
after being caught. The catch displays "NO!" on the calculator. Need to
change it so that pressing other buttons clears the text first.

Pressing equal while the expression is not whole is supposed to do
nothing, but right now it returns "NaN". To fix.